tdf#128736 sw ContinuousEndnotes: fix use-after-free on text frame join
Regression from commit 61cf196631a2a846e0d3b8b83c0805cf4d1d14b2 (sw
ContinuousEndnotes: fix moving them to the previous page, 2019-10-25),
the problem was that the SwFootnoteFrame::mpReference was not updated on
frame split. This meant that by the time the frame was joined,
SwTextFrame::JoinFrame() thought that the follow has no footnotes, so
the footnote reference was not updated, resulting in a dangling pointer.

Fix the problem by going back to using bEnd for endnotes (both the Word
compat and the normal case), this means that
SwTextFrame::ConnectFootnote() will invoke
SwFootnoteBossFrame::ChangeFootnoteRef(), fixing the dangling pointer.

Then fix the original problem differently: similar to the in-section
endnotes, just remove + append them, this makes sure that the endnotes
(in the Word compat case) move to a previous page on page delete.

Change-Id: Ia1b8e54b4a0b0f385c703f8e7016011c3ac57a03
Reviewed-on: https://gerrit.libreoffice.org/82778
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins