tdf#105481 : Do not skip invalidation of as-char anchored objects
inside a growing text-frame after doing recalc of positions of
other frames (visually above current) in the tree. Otherwise
the as-char anchored objects gets wrong reference (base) position
(more accurately the base Y position) and hence get displayed at
wrong vertical positions (in case of the bug document, it goes
out of the page area).

Some notes about the bug:

Above mentioned problem is visible (at least) when you have a frame
anchored to the bottom of the page with auto-height and it contains
one or more <text:p> with at least one of them having a draw object
like a shape or math formula with 'as-char' anchoring. Only the
draw object in the last <text:p> is affected by the bug possibly
because this text-frame is the cause of growth of the parent (upper)
frame.

Change-Id: If968c8c00aa57d26b1000e3250b352b24df47cf6
Reviewed-on: https://gerrit.libreoffice.org/83603
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>