Correctly set the macro security auto confirm mode
The code of adjustMacroMode() has one major problem: (almost)
everything is handled by the single nMacroExecutionMode variable,
which is even changed based on its initial setting.

I don't know if this fixed functionality breaks something, but it
doesn't look like it, as the enum AutoConfirmation is just used
and defined locally.

Change-Id: Iac305e9ec0020a11fd53b4bb87cc782686a247ac
Reviewed-on: https://gerrit.libreoffice.org/85055
Tested-by: Jenkins
Reviewed-by: Jan-Marek Glogowski <glogow@fbihome.de>