make Skia Windows widget drawing use correct alpha (tdf#129074)
The OpenGL code made the widget drawing use two variants drawn
on black and on white, for reasons not explained
in 3149cc341b1866d215110f0783227549a99b5920 (probably the Windows
API doesn't handle alpha completely correctly or whatever).
This means that getting the actual alpha requires a custom
algorithm that needs to be implemented manually for Skia use.

Change-Id: I1438f3829a1bdeda9e55700c4a397c60d5663446
Reviewed-on: https://gerrit.libreoffice.org/84948
Tested-by: Jenkins
Reviewed-by: Luboš Luňák <l.lunak@collabora.com>