tdf#129522 ww8import/export: allow char shadow_NONE overrides
An earlier patch did this for paragraphs. The same is needed
for character properties.

This also applies to borders - where a non-border needs to be
processed in order to override a style-defined border.

This export change doesn't affect DOCX immediately, because
DOCX will do nothing with a NONE-border/shadow until style
information is provided to it. So this change is safe to make
on it's own as far as DOCX is concerned.

RTF just exports the border, but since it doesn't do any of
this well, add in the logic so that RTF is unaffected...

Change-Id: Ia857dc0db7d7efc83145a14e4fe4052d32b539ac
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/85760
Tested-by: Jenkins
Reviewed-by: Justin Luth <justin_luth@sil.org>