cleanup tdf#129522: indenting
It is much easier to see what has changed if a commit does not
include formatting changes. Since there were a couple of them,
I separated it out into a separate patch.

I don't normally do this, but it seems like a good idea,
so I'm trying out the best way to do it.
The bad-side would be an extra patch to backport,
but I'm not expecting much desire to backport here.

Change-Id: Icc0eae1a881af80c0b8230697b235d1901b55fe3
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/85871
Tested-by: Jenkins
Reviewed-by: Justin Luth <justin_luth@sil.org>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>