cid#1456609 Operands don't affect result
2.2.2.37 EmfPlusPointR Object has...

"X (variable) This value MUST be specified by either an EmfPlusInteger7 object
or an EmfPlusInteger15 object." and the same for Y. where variable is variable
length.

I think ReadInt32 isn't the right choice here and we need to read either one or
two bytes, using the highbit of the first byte to determine if we need to
read another byte.

Change-Id: I60c0687403ff58dc393bd55a22f37c89357f60c3
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/86207
Tested-by: Jenkins
Reviewed-by: Caolán McNamara <caolanm@redhat.com>
Tested-by: Caolán McNamara <caolanm@redhat.com>