tdf#129798: store SwWrtShell when it's available
GetActiveWrtShell may return nullptr when attaching a new component
to current frame, e.g. destroying on print preview. OTOH SwWrtShell
is passed to SwFieldDBPage::SetWrtShell on creation, so store it to
own base class then.

Possibly that would be enough, and we could remove all the checks
in other methods - but I don't know the code well enough to be sure
that SwFieldDlg::PageCreated and SwFieldEditDlg::CreatePage will
always set SwWrtShell; to be on the safe side, I kept the checks
where SwWrtShell is used, and stored it in those places if not yet
stored.

Regression after commit 39ba5e3bde93358af1b363da8f850bdc96806cfa.

Change-Id: I21f3c8ef24764e98b1bc287db3c6485ce0b3ab48
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/86253
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>